RawSubscripts correspond to GEMM and DOT
